requirements.txt
django==5.2.7
djangorestframework==3.16.1
requests==2.32.5
gitpython==3.1.45
dotenv==0.9.9
python-dotenv~=1.1.1
django-jazzmin~=3.0.1